Add support for "Go To Definition" in partial classes (Issue #53)
Currently SpecFlow only supports the "Go To Definition" for full(?) classes (non-partial classes) because it only looks at classes which are marked with the [Binding] attribute. This change extends the search to partial classes by first finding all classes marked with [Binding] then checking to see whether any of these are partial classes. If so, those partial classes are assumed to be binding classes and are scanned as well.