Whitespace change indicate location of threading issue
empty change to indicate location of threading issue.

When a Solution is closed, this code path is run on the UI via the event handler for OnSolutionClosed (from ProjectScopeFactory). If there are background threads running, and they try and access the UI thread (for example, if a background thread is attempting to access some VS Project COM API, it will be marshalled onto the UI thread), then this could result in a deadlock. The UI thread is blocked waiting for background threads to complete, but background threads are blocked on access to UI thread.

Have to be careful that no background thread is accessing the UI thread when this lock is held (or maybe look at a way to handle the OnSolutionClosedEvent without blocking the UI thread).